requirements.txt
pyautogui
pillow
requests
